Bug Fixes
BUGS:

1) in ObsTemplate._set_prior:
The pbm with factory=datamodel.get is that when initialized a KeyModel cretaes a dummy player for which it gets a rating, which is passed to the datamodel of the factory that acts as a default dict -> this dummy player is added to the original datamodel leading to massif side-effect in the simulation.
FIX:
+ prior is the datamodel (which seems obvious ...)
+ HACK fix for posterior, may need changes

2) Elo OverFlowError
The pbm was in the rating2-rating1 and the way it is passed to the logistic_elo func
FIX:
+ test/inferer/test_elo: the new_ranking Elo implementation passes test, the one in rstt.ranking does not !

ALSO:
the new_ranking implementation reproduces the master thesis experiments and results.